requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.10.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
Faker==9.8.0
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.2
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
python-twitter==3.5
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
